+++ src/RoadCaptain.RouteBuilder/WindowService.cs
@@ -1,4 +1,6 @@
-﻿using Autofac;
+﻿using System;
+using Autofac;
+using Microsoft.Win32;
 using RoadCaptain.UserInterface.Shared;
 
 namespace RoadCaptain.RouteBuilder
@@ -26,6 +28,27 @@ namespace RoadCaptain.RouteBuilder
 
                 Show(window);
             }
+        }
+
+        public string ShowSaveFileDialog()
+        {
+            var dialog = new SaveFileDialog
+            {
+                RestoreDirectory = true,
+                AddExtension = true,
+                DefaultExt = ".json",
+                Filter = "JSON files (.json)|*.json|GPS Exchange Format (.gpx)|*.gpx",
+                InitialDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments)
+            };
+
+            var result = ShowDialog(dialog).GetValueOrDefault();
+
+            if (!result)
+            {
+                return null;
+            }
+
+            return dialog.FileName;
         }
     }
 }