+++ Changelog.md
@@ -4,6 +4,10 @@
 
 - When starting, both the runner and route builder will now check if there is a new release available and shows a window to inform you if there is: ![Screenshot of a window showing the new version information](./images/runner-new-version.png) ([issue #15](https://github.com/sandermvanvliet/RoadCaptain/issues/15))
 
+### Route Builder
+
+- Dialogs should now appear in or over the main window instead of outside of it. ([issue #26](https://github.com/sandermvanvliet/RoadCaptain/issues/26))
+
 ## 0.5.4.0
 
 ### Runner